Refactor JS for better structure and maintainability
- Move data-related functions (`getCurrentStreak`, `getPunkteFuerTag`) from `ui.js` to `tasks.js`.
- Create a new `effects.js` module for UI effects (`starteKonfetti`, `createRipple`) and move them from `ui.js`.
- Add a missing `addDays` utility function to `utils.js` to fix a bug in the task modal.
- Remove the unused `escapeHTML` function from `utils.js`.
- Remove a redundant call to `updateMetaBar` in `ui.js`.
- Update all relevant imports to reflect the new file structure.
